Add files via upload
Title: Predicting Canada's Per Capita Income in 2020


This repository contains code for a machine learning project aimed at predicting Canada's per capita income for the year 2020. Leveraging historical data spanning from 1978 to 2016, provided in a CSV format, our goal is to develop a robust predictive model that can accurately forecast the per capita income for Canada in 2020.

Key Features:
- Data Preprocessing: Comprehensive preprocessing steps including cleaning, handling missing values, and feature engineering to prepare the dataset for model training.
- Exploratory Data Analysis (EDA): In-depth exploratory analysis to gain insights into the underlying patterns and trends in the data.
- Model Development: Implementation of various machine learning algorithms such as linear regression, decision trees, random forests, and gradient boosting to build predictive models.
- Model Evaluation: Rigorous evaluation of model performance using appropriate metrics such as mean squared error, R-squared, and cross-validation to ensure reliability and accuracy.
- Prediction for 2020: Application of the trained model to predict Canada's per capita income for the year 2020 based on the available historical data.
- Documentation: Detailed documentation including explanations of methodologies, code comments, and insights derived from the analysis.

Dependencies:
- Python 3.x
- Pandas
- NumPy
- Scikit-learn
- Matplotlib
- Seaborn

Contributing:
Contributions to the project are welcome! Whether it's code improvements, bug fixes, or feature enhancements, feel free to fork the repository and submit pull requests.



Acknowledgments:
I would like to express our gratitude to the contributors of the dataset used in this project. Additionally, thanks to the open-source community for the valuable libraries and tools utilized in this project.